Rename some auth/ components
